Add files via upload
The goal of this project is to develop a trading algorithm for Bank of America's stock that utilizes various classification and regression analysis techniques to predict future price movements. To achieve this goal, historical data on Bank of America's stock prices, as well as relevant economic and financial indicators, will be collected and analyzed.

First, classification analysis techniques will be used to identify different market trends and patterns that can be used to predict the future movement of Bank of America's stock. This will include techniques such as clustering, decision trees, and random forests.

Next, regression analysis techniques will be used to create a predictive model that takes into account various market and economic factors, such as interest rates, inflation, and GDP growth, that could impact the price of Bank of America's stock. These factors will be used to build a time-series model that can be used to forecast future price movements.

Once these models have been developed and tested, they will be incorporated into a trading algorithm that will buy and sell Bank of America's stock based on their predictions. The algorithm will be back-tested using historical data to ensure that it performs well under a variety of market conditions.

Overall, this project aims to develop a robust and effective trading algorithm for Bank of America's stock that leverages the power of classification and regression analysis to predict future price movements and generate profits for investors.